+++ src/Services/MPesaMz.php
@@ -13,7 +13,7 @@ use abdulmueid\mpesa\Transaction;
  * @since 0.1.0
  * @see Transaction
  */
-class MpesaMz
+class MPesaMz
 {
     /**
      * Class that defines all the M-Pesa API transactions
@@ -23,21 +23,31 @@ class MpesaMz
     protected $transaction;
 
     /**
+     * Path of the config file of M-Pesa API
+     *
+     * @var string
+     */
+    protected $configPath;
+
+    /**
      * Create a new instance
      *
+     * @param  string  $configPath
      * @return void
      */
-    public function __construct()
+    public function __construct(string $configPath = null)
     {
+        $this->configPath = $configPath ?? config_path('mpesa-config.php');
+
         $this->transaction = new Transaction(
-            Config::loadFromFile(config_path('mpesa-config.php')));
+            Config::loadFromFile($this->configPath));
     }
 
     /**
      * Forwards the calls to transaction
      *
-     * @param string $name
-     * @param array $arguments
+     * @param  string  $name
+     * @param  array  $arguments
      *
      * @return TransactionResponseInterface
      */